BlindSpot algorithm
It really doesn't perform as well as I expected, this was geared to fit
the test case b.txt.

RocketBoostedEvolution was a fun one though. A direct copy of
Evolutionary, but with the beginning path generated by MergeIterative.
Solves TS like a boss, I'm currently using it as the "correct" form of
anything in the 10-100 range (those are too large to brute force)